Added XML request and response support.
This adds 'xml' and 'xmlData' mirroring 'json' and 'jsonData' to scotty.
Unfortunately, this also introduces a dependency on xml-conduit.
You might ask who uses XML for REST nowadays anyway. That is a good point,
but should scotty ever want XML support, here it is.

I also added a minimalistic example on XML consumption and production.