Remove the `runM` parameter to `scotty*T`
`ScottyT` and `ActionT` shared the same monad as a type parameters, which makes
it inconvenient to use a monad stack in routes since you then have to use the
same stack during initialization – often without them having anything to do
with eachother.

Initialization can be done before calling `scotty*T`, so there is no
need to have this one-time action affect the monad parameter of
actions.