Make ScottyT a transformer again
We can have decoupled monads for ScottyT and ActionT if ScottyT is parametrised
by two monads. This breaks the API a lot because it changes the kind of ScottyT
but it allows for composable components.
With this change it is possible to write several "components" of type `ScottyT e
m n a` and bind them together. By a "component" I mean something that can do
its own setup and teardown in some custom monad stack (e.g. open db connection
in IO).
This is currently not possible since ScottyT is not a true transformer but fixed
over identity.