Add MonadReader and MonadState ActionT instances (#232)
It's probably a fairly common case to have a MonadReader as 'm' in
'ActionT e m' which stores some global configuration. Accessing this
reader can only be done with explicit 'lift'.

This commit add 'MonadReader (ActionT e m)' instance which is forwarded
from 'm' itself, to avoid this 'lift' and simplify user code. The same
is done for 'MonadState'.