Allow building with text-2.0
For the most part, this is simply a matter of raising the upper version bounds
in the `.cabal` file. The max-size request test case in the test suite needed
to be adjusted, as it needs to pick a sufficiently large request such that it
takes multiple chunks to store it as a lazy Text value. With `text-2.0` or
later, the amount of ASCII characters that a single chunk can store is much
larger due to using UTF-8 encoding internally, so we must generate a larger
request to compensate.